semantic_search/requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
markdown>=3.5
